Better handling of EntityKind SQS/SNS transport (#377)
Add support for using both `Queue` and `Broadcast` entity kinds in the AmazonSqs transport. The EntityKind set on the EventRegistration is respected. A topic is only created for Broadcast. As a result, bulk publish and delay (scheduled) are now supported when using EntityKind.Queue since SQS supports it but SNS does not. When used with EntityKind.Broadcast, a warning is produced.