Access raw transport data in serializers (#381)
This PR adds support for sending the raw data received from the transport into the `DeserializationContext` so that it is available to the serializer in case more checks need to be done. Further, the `DeserializerToEnvelope<T>(...)` method now takes `DeserializationContext` instead of `ContentType` so that more information is available downstream.