Samples for multiple transports (#457)
This PR adds samples on how to work with multiple transports of the same type or different types. For the same type, the InMemory transport is used for quick testing. For different types, both Azure Service Bus and Azure Event Hubs (for IoTHub) is used which is a common scenario.