Transports should be created each time especially for different types (#459)
When multiple transports of the same type are registered, each will need to start consuming on its own. When resolved from the service provider, only one instance is created. Instead, a new instance should be created for each registered transport.

This PR solves that and removes registration of transport instances to the IoC container.