Consuming for dead-lettered events (#491)
Support consuming events from dead-letter entities. This works best for Azure Service Bus that has the concept of a dead-letter sub-queue. For other transports, this is supported using a separate dead-letter entity that is usually named `{entity-name}-deadletter`. These are mainly meant for simple scenarios where the handler cannot error because we cannot dead-letter a dead-lettered event.