AOT compatibility (#621)
This replaces `IsTrimmable` with `IsAotCompatible` which contains better analyzers. As a result, the deserialize and consume logic has been updated to handle specifics such as generics that cannot be inferred in AOT.

Initial support for trimming in #564 did not have support for JSON serializer. Using the slim event bus meant you needed to create your own serializer and register it as a default. With this PR, the `DefaultJsonEventSerializer` support the use of a `JsonSerializerContext` where each event has been declared as `EventEnvelope<TEvent>`.

Consequently, registration of consumers requires explicit events as shown in the sample